requirements.txt
<<<<<<< HEAD
streamlit
torch
torchvision
numpy
pillow
opencv-python-headless
timm
=======
streamlit
torch
torchvision
numpy
pillow
opencv-python-headless
timm
>>>>>>> 079c569 (Streamlit Cloud ready)
